Move functions to separate files
