+++ starknet_py/cairo/selector.py
@@ -1,0 +1,16 @@
+from starknet_py.cairo.utils import _starknet_keccak
+from starknet_py.constants import (
+    DEFAULT_ENTRY_POINT_NAME,
+    DEFAULT_L1_ENTRY_POINT_NAME,
+    DEFAULT_ENTRY_POINT_SELECTOR,
+)
+
+
+def get_selector_from_name(func_name: str) -> int:
+    """
+    Returns the selector of a contract's function name.
+    """
+    if func_name in [DEFAULT_ENTRY_POINT_NAME, DEFAULT_L1_ENTRY_POINT_NAME]:
+        return DEFAULT_ENTRY_POINT_SELECTOR
+
+    return _starknet_keccak(data=func_name.encode("ascii"))
