+++ starknet_py/net/account/account_client.py
@@ -4,10 +4,8 @@ import warnings
 from dataclasses import replace
 from typing import Optional, List, Union, Dict, Tuple, Iterable
 
-from starknet_py.common import (
-    create_compiled_contract,
-    get_selector_from_name,
-)
+from starknet_py.cairo.selector import get_selector_from_name
+from starknet_py.common import create_compiled_contract
 from starknet_py.compile.compiler import StarknetCompilationSource
 from starknet_py.constants import FEE_CONTRACT_ADDRESS
 from starknet_py.net.account.compiled_account_contract import COMPILED_ACCOUNT_CONTRACT
