+++ starknet_py/net/udc_deployer/deployer.py
@@ -4,7 +4,8 @@ from typing import Optional, List, Union, cast, NamedTuple
 
 from starkware.starknet.definitions.fields import ContractAddressSalt
 
-from starknet_py.common import int_from_hex, get_selector_from_name
+from starknet_py.cairo.selector import get_selector_from_name
+from starknet_py.common import int_from_hex
 from starknet_py.constants import DEFAULT_DEPLOYER_ADDRESS
 from starknet_py.net.client_models import Hash, Call
 from starknet_py.net.models import AddressRepresentation, parse_address, compute_address
