+++ starknet_py/proxy/proxy_check.py
@@ -2,7 +2,8 @@ import re
 from abc import ABC, abstractmethod
 from typing import Optional, Callable
 
-from starknet_py.common import get_storage_var_address, get_selector_from_name
+from starknet_py.cairo.selector import get_selector_from_name
+from starknet_py.cairo.storage import get_storage_var_address
 from starknet_py.constants import RPC_INVALID_MESSAGE_SELECTOR_ERROR
 from starknet_py.net.client import Client
 from starknet_py.net.client_errors import ClientError
