+++ starknet_py/tests/e2e/fixtures/proxy.py
@@ -3,7 +3,7 @@
 import pytest
 import pytest_asyncio
 
-from starknet_py.common import get_selector_from_name
+from starknet_py.cairo.selector import get_selector_from_name
 from starknet_py.contract import Contract, DeployResult
 from starknet_py.net import AccountClient
 from starknet_py.tests.e2e.fixtures.constants import CONTRACTS_PRECOMPILED_DIR, MAX_FEE
