+++ starknet_py/utils/typed_data.py
@@ -5,7 +5,7 @@ from marshmallow import Schema, fields, post_load
 from starkware.cairo.common.hash_state import compute_hash_on_elements
 
 from starknet_py.cairo.felt import encode_shortstring
-from starknet_py.common import get_selector_from_name
+from starknet_py.cairo.selector import get_selector_from_name
 from starknet_py.net.models.typed_data import StarkNetDomain
 from starknet_py.net.models.typed_data import TypedData as TypedDataDict
